dispatch-app/src/main/java/org/chamara/dispatchapp/service/DispatchService.java
package org.chamara.dispatchapp.service;

import lombok.RequiredArgsConstructor;
import org.chamara.dispatchapp.message.OrderCreated;
import org.chamara.dispatchapp.message.OrderDispatched;
import org.springframework.kafka.core.KafkaTemplate;
import org.springframework.stereotype.Service;

import java.util.concurrent.ExecutionException;

@Service
@RequiredArgsConstructor
public class DispatchService {
    private final static String ORDER_DISPATCHED_TOPIC = "order.dispatched";
    private final KafkaTemplate<String, Object> kafkaProducer;

    public void process(OrderCreated payload) throws Exception {
//        no op method
        OrderDispatched orderDispatched = OrderDispatched.builder()
                .orderId(payload.getOrderId())
                .build();
        kafkaProducer.send(ORDER_DISPATCHED_TOPIC, orderDispatched).get();
    }

}
